AI_Cloud_Assistant/requirements.txt
# Core data
pandas
numpy
matplotlib
scikit-learn

# LLMs / Agents
google-generativeai
google-adk
litellm

# Env
python-dotenv

# LangChain stack
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-groq
langchain-chroma

# RAG / Embeddings
sentence-transformers
chromadb
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
pypdf
docx2txt

# Forecasting (your existing)
prophet
cmdstanpy

# Streamlit UI  
streamlit>=1.32.0

# Utilities
requests
tenacity

# Validation
pydantic
